% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{badOption}
\alias{badOption}
\title{bad argument error handling function}
\usage{
badOption(arg)
}
\arguments{
\item{arg}{argument choosen}
}
\description{
bad argument error handling function
}
\keyword{internal}
