% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{c19.genomic.data}
\alias{c19.genomic.data}
\title{function to obtain genomic data from SARS-CoV-2019}
\usage{
c19.genomic.data(src = "livedata", accOnly = TRUE)
}
\arguments{
\item{src}{argument to indicate what sources are going to be used for retrieving the data: "livedata", "repo" or "local"
'livedata' will access NCBI servers to acquire the latest possible data, this may incur in significant longer times
  'repo' will access an updated replica of the data from a github repository (faster but not necessarily upto the latest udpates)
  'local' will access previously archived records within teh package (fastest but not updated)}

\item{accOnly}{boolean indicator for getting only accession codes or whole records}
}
\value{
a list containing reference genome, annotation data, nucleotides, proteins and list of SRA runs
}
\description{
function to obtain genomic data from SARS-CoV-2019
}
