% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{c19.NP_fasta.data}
\alias{c19.NP_fasta.data}
\title{function to obtain FASTA seqs for nucleotides or proteins from SARS-CoV-2}
\usage{
c19.NP_fasta.data(src = "repo", target = "nucleotide")
}
\arguments{
\item{src}{origin for the data source: "livedata" OR "repo"}

\item{target}{"nucleotide", "protein" or "codingRegion"}
}
\description{
function to obtain FASTA seqs for nucleotides or proteins from SARS-CoV-2
}
