% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{c19.NPs.data}
\alias{c19.NPs.data}
\title{function to obtain data for nucleotides or proteins from SARS-CoV-2}
\usage{
c19.NPs.data(
  src = "livedata",
  DB = "nucleotide",
  max.nr.recs = NULL,
  accOnly = TRUE
)
}
\arguments{
\item{src}{origin for the data source: "livedata", "repo", "local"}

\item{DB}{database}

\item{max.nr.recs}{maximun number of records to retrieve, there are limitations in the fns and server sides}

\item{accOnly}{boolean indicator for getting only accession codes or whole records}
}
\description{
function to obtain data for nucleotides or proteins from SARS-CoV-2
}
