% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{c19.refGenome.data}
\alias{c19.refGenome.data}
\title{function to obtain sequencing data grom NCBI
Reference:  https://www.ncbi.nlm.nih.gov/nuccore/NC_045512.2}
\usage{
c19.refGenome.data(src = "livedata", graphics.ON = TRUE)
}
\arguments{
\item{src}{data origin source: 'livedata', 'repo', 'local'}

\item{graphics.ON}{flag to activate/deactivate graphical output}
}
\description{
function to obtain sequencing data grom NCBI
Reference:  https://www.ncbi.nlm.nih.gov/nuccore/NC_045512.2
}
\examples{
# obtain covid19's genomic data
covid19.gen.seq <- c19.refGenome.data()
# display the actual RNA seq
covid19.gen.seq$NC_045512.2

}
