% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_Canada.R
\name{covid19.URL_csv.data}
\alias{covid19.URL_csv.data}
\title{function to read CSV from URLs or local replicas}
\usage{
covid19.URL_csv.data(
  local.data = FALSE,
  acknowledge = FALSE,
  srcURL = "",
  srcName = "",
  locFileName = NA,
  locVarName = NA
)
}
\arguments{
\item{local.data}{boolean flag to indicate whether the data will be read from the local repo, in case of connectivity issues or data integrity}

\item{acknowledge}{boolean flag to indicate that the user acknowledges where the data is coming from.  If FALSE, display data acquisition messages.}

\item{srcURL}{URL from where to obtain the data}

\item{srcName}{name of the source}

\item{locFileName}{name of the file to read from local repo}

\item{locVarName}{name of the variable loaded from local file}
}
\value{
data as oriignally obtained from the URL src
}
\description{
function to read CSV from URLs or local replicas
}
