% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_vaccination_testing.R
\name{covid19.vaccination}
\alias{covid19.vaccination}
\title{function to read data related to covid19 vaccinations}
\usage{
covid19.vaccination(tgt = "global", data.fmt = "orig", disclaimer = TRUE)
}
\arguments{
\item{tgt}{selects data type: 'global','us','country','locations'}

\item{data.fmt}{selects the format of the data, options are: 'orig' (original as reported by the OWID repo)}

\item{disclaimer}{indicates whether the information about the source of the data is disclosed}
}
\description{
function to read data related to covid19 vaccinations
}
