% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_genomic.R
\name{getFile}
\alias{getFile}
\title{auxiliary function to download files in a protected fashion, i.e. against
errors in the downloading procedure}
\usage{
getFile(url = NULL, fileName = NULL)
}
\arguments{
\item{url}{resource's URL on the web}

\item{fileName}{name of the resource (file) to download

keywords internal}
}
\description{
auxiliary function to download files in a protected fashion, i.e. against
errors in the downloading procedure
}
