% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_utils.R
\name{header0}
\alias{header0}
\title{auxiliary fn to print "headers" adn 'titles'}
\usage{
header0(x = "-", title = "", total.len = 80, eol = "\\n")
}
\arguments{
\item{x}{character to use as lines}

\item{title}{title to dispo}

\item{total.len}{length of the line}

\item{eol}{end of line character}
}
\description{
auxiliary fn to print "headers" adn 'titles'
}
\keyword{internal}
