% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_rollingEstimates.R
\name{mrollingRates}
\alias{mrollingRates}
\title{function to compute a rolling fn (rate) of multiple quantities from TS data, eg. fatality and recovery rate}
\usage{
mrollingRates(data = NULL, geo.loc = NULL, fn = mean, period)
}
\arguments{
\item{data}{time series dataset to consider}

\item{geo.loc}{country/region to analyze}

\item{fn}{function to compute rolling}

\item{period}{length of window}
}
\description{
function to compute a rolling fn (rate) of multiple quantities from TS data, eg. fatality and recovery rate
}
