% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_anal.R
\name{mtrends}
\alias{mtrends}
\title{function to visualize different indicators for trends in daily changes of cases reported as time series data, for mutliple (or single) locations}
\usage{
mtrends(data, geo.loc = NULL, confBnd = TRUE, info = "")
}
\arguments{
\item{data}{data.frame with *time series* data from covid19}

\item{geo.loc}{list of locations}

\item{confBnd}{flag to activate/deactivate drawing of confidence bands base on a moving average window}

\item{info}{additional info to display in the plot}
}
\description{
function to visualize different indicators for trends in daily changes of cases reported as time series data, for mutliple (or single) locations
}
\examples{
ts.data <- covid19.data("ts-confirmed")
mtrends(ts.data, geo.loc=c("Canada","Ontario","Uruguay","Italy"))

}
