% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_checks.R
\name{nullify.data}
\alias{nullify.data}
\title{remove inconsistencies from data by removing 'suspicious' entries}
\usage{
nullify.data(data, stringent = FALSE)
}
\arguments{
\item{data}{dataset to process}

\item{stringent}{only return records with "complete cases"}
}
\description{
remove inconsistencies from data by removing 'suspicious' entries
}
