% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_vaccination_testing.R
\name{OWID.repos}
\alias{OWID.repos}
\title{function to define the OWID repos URLs}
\usage{
OWID.repos(tgt)
}
\arguments{
\item{tgt}{target case:  'VAC.global','VAC.us','VAC.country','VAC.locations', 'testing', 'testing.details'}
}
\value{
URL
}
\description{
function to define the OWID repos URLs
}
