% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_models.R
\name{plot.SIR.model}
\alias{plot.SIR.model}
\title{function to plot the results from the SIR model fn}
\usage{
\method{plot}{SIR.model}(SIR.model, geo.loc = "", interactiveFig = FALSE, fileName = NULL)
}
\arguments{
\item{SIR.model}{model resulting from the generate.SIR.model() fn}

\item{geo.loc}{optional string to specify geographical location}

\item{interactiveFig}{optional flag to activate interactive plot}

\item{fileName}{file where to save the HTML version of the interactive figure}
}
\description{
function to plot the results from the SIR model fn
}
\keyword{internal}
