% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_utils.R
\name{preProcessingData}
\alias{preProcessingData}
\title{auxiliary function to pre-process data per geographical location}
\usage{
preProcessingData(data0, geo.loc)
}
\arguments{
\item{data0}{data set}

\item{geo.loc}{geopgraphical location, can be a country, region, province or city}
}
\description{
auxiliary function to pre-process data per geographical location
}
\keyword{internal}
