% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_print.R
\name{print.report}
\alias{print.report}
\title{method associated with 'report' objects}
\usage{
\method{print}{report}(report.obj)
}
\arguments{
\item{report.obj}{'report' object}
}
\description{
method associated with 'report' objects
}
\keyword{'report'}
\keyword{associated}
\keyword{function}
\keyword{internal}
\keyword{method}
\keyword{object}
\keyword{the}
\keyword{to}
