% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_rollingEstimates.R
\name{rollingRate}
\alias{rollingRate}
\title{function to compute a rolling fn of a TS data}
\usage{
rollingRate(data, fn = mean, period = NULL)
}
\arguments{
\item{data}{TS data}

\item{fn}{function to compute rolling}

\item{period}{length of window}
}
\value{
a vector with rolling values
}
\description{
function to compute a rolling fn of a TS data
}
