% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_utils.R
\name{set.plt.canvas}
\alias{set.plt.canvas}
\title{auxiliary function to set the graphical layout}
\usage{
set.plt.canvas(geo.loc, ylayers = 1, minBreaks = 5)
}
\arguments{
\item{geo.loc}{list of locations so that the fn can determine how many plots would be}

\item{ylayers}{parameter to set a multiplier times the nbr of initial plots}

\item{minBreaks}{cut-off to set the minimum nbr of plots}
}
\description{
auxiliary function to set the graphical layout
}
\keyword{internal}
