% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_models.R
\name{simple.SIR.ODE}
\alias{simple.SIR.ODE}
\title{Define an ODE system for a simple SIR model}
\usage{
simple.SIR.ODE(time, state, parameters)
}
\arguments{
\item{time}{time variable}

\item{state}{state variable}

\item{parameters}{parameters of the ODE}
}
\description{
Define an ODE system for a simple SIR model
}
\keyword{internal}
