% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_seq.R
\name{X.covid19.genomic.data}
\alias{X.covid19.genomic.data}
\title{function to obtain sequencing data grom NCBI
Reference:  https://www.ncbi.nlm.nih.gov/nuccore/NC_045512.2}
\usage{
X.covid19.genomic.data(graphics.ON = TRUE)
}
\arguments{
\item{graphics.ON}{flag to activate/deactivate graphical output}
}
\description{
function to obtain sequencing data grom NCBI
Reference:  https://www.ncbi.nlm.nih.gov/nuccore/NC_045512.2
}
\examples{
# obtain covid19's genomic data
covid19.gen.seq <- covid19.genomic.data()
# display the actual RNA seq
covid19.gen.seq$NC_045512.2

}
